Remove red/green on death saving throws and make it ignore advantage roll mode
The red/green for success/failure caused confusion to some players who thought
it meant a critical success/failure, and the formula makes it even more confusion
to those not familiar with R20 dice formula syntax, especially in a 'life or death'
stressful situation.
Remove advantage rolls for death saves because the roll twice would be irrelevant
here most of the time and the user can double click if needed.