FVTT: Update CSS to make it look nicer and fix tooltips.
The CSS for the 'roll cells' was provided by @solfolango77 and it
makes it look so much nicer! The trick to make the tooltip position
correct was to make it's closest parent with a 'position' be the one that
fits the entire space (roll-result) and to use top/left to set it correctly.
Also removes a debugging for buttons which I forgot was there.