Added setting for dex score initiative tiebreaker
Implemented by accumulating the dexterity score of a monster if it is defined and adding it as a decimal. For some JavaScript calculations, this results in floating point accuracy edge cases, so we round off two decimals to ensure our result is rendered properly by the default Beyond20 & VTT renderer.