Change native rolls to use real actor and do things a bit differently
This is still a WIP but it now requires a real actor to be present, which the GM can create
or the player if they have authorization.
There are still many issues it seems with how Foundry works as it seems to do everything it can to interfere
with the data being set on the actor.