Add support for whispering to gamelog
Add request when resolving rolls and add whisper option to the digital dice
roll itself, which allows us to send to DM or send to Self (or fallback to public)
on whispered rolls.
We also preserve the entity type/id, gameId/userId and message scope/target
of digital dice rolls when intercepting the message broker messages.
Fixes #908