Prevent double injection of FVTT script
Injecting the fvtt_test script caused the tab to get a new update which made
it re-inject it as it 'finished loading'. This caused console errors or variables
being re-defined. By making sure we only inject in tabs that aren't already
registered, and registering the tabs immediately on icon activation, we avoid
double-loading of the scripts.