Remove forge-vtt from manifest content script
The script still gets injected with the new FVTT auto inject code, but having
it defined in the manifest will cause it to be injected twice, once by the background
script and once by the browser itself due to the manifest.
As long as the forge-vtt.com domain remains in the explicit permissions,
it should not have any visible impact on the behavior.