Add limited support for Goliath "Large Form" trait
This commit adds a partial implementation of the racial trait called "Large Form" that has been introduced for Goliaths as of the latest One D&D playtest. The effects of this bonus action are to (1) grant ADV on STR checks for the duration of the effect, (2) increase the size of the character to Large, and (3) increase the speed of the character by 10 feet. Only the first of these is implemented because it is unclear whether it is even possible (let alone advisable) to temporarily bump the size and speed of the character.